Revert "Mapping and IHttpActionResult Methods"
This reverts commit 3cd0bde0b0f60bf2250e8ef42268220d54746458.

+++ ProductProject/Models/Genre.cs
@@ -8,7 +8,8 @@ namespace ProductProject.Models
 {
     public class Genre
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
+
         [Required]
         [StringLength(255)]
         public string Name { get; set; }
